enforce wrapper and transformer on same device
This was covered up by previous usage where we'd instantiate trainers for each transformer (semantic/coarse/fine). Now that we have to be able to load transformers without their corresponding trainers, because we can only load one accelerator at a time, the code in trainer that checks wrapper device no longer applies and we get device mismatches. This commit fixes that